+++ price-feeder/CHANGELOG.md
@@ -52,6 +52,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - [1328](https://github.com/umee-network/umee/pull/1328) Add bitget provider.
 - [1339](https://github.com/umee-network/umee/pull/1339) Add mexc provider.
+- [1430](https://github.com/umee-network/umee/pull/1430) Support 'test' keyring-backend.
 
 ### Bugs